requirements.txt
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Config / env
python-dotenv>=1.0.0

# RSS
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Telegram bot
python-telegram-bot>=21.0
